requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
werkzeug
python-dotenv